config/requirements.txt
# MCPVotsAGI Jupiter Integration Requirements
# ==========================================

# Core Dependencies
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
aiofiles>=23.0.0
websockets>=11.0.0
psutil>=5.9.0
numpy>=1.24.0
pandas>=2.1.0
pyyaml>=6.0.1
requests>=2.31.0

# Enhanced A2A Communication System
redis>=4.5.0
aioredis>=2.0.0
msgpack>=1.0.0
aiodns>=3.0.0

# Jupiter DEX Integration
solana>=0.30.0
anchorpy>=0.18.0
construct>=2.10.0

# Claude/AI Integration
anthropic>=0.7.0
openai>=1.3.0
tiktoken>=0.5.0
ipfshttpclient>=0.8.0
ollama>=0.1.0

# Trading Dependencies
python-binance>=1.0.0
yfinance>=0.2.0
ccxt>=4.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Machine Learning / RL
tensorflow>=2.14.0
torch>=2.1.0
stable-baselines3>=2.1.0
gym>=0.29.0

# Monitoring & Logging
rich>=13.6.0
colorama>=0.4.6

# Web Scraping and Automation
selenium>=4.0.0
beautifulsoup4>=4.11.0
playwright>=1.30.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.0.0
isort>=5.12.0

# Time & Date
pytz>=2023.3
python-dateutil>=2.8.2

# Configuration
pydantic>=2.4.0
python-dotenv>=1.0.0

# Blockchain & Crypto
web3>=6.11.0
eth-account>=0.9.0
cryptography>=41.0.0

# System Integration
schedule>=1.2.0
watchdog>=3.0.0
click>=8.1.7

# Claude Code Usage Monitor
typer>=0.9.0
typing-extensions>=4.8.0

# DGM Integration Dependencies
prometheus-client>=0.19.0
aiocache>=0.12.0
tenacity>=8.2.0

# Optional Dependencies
jupyter>=1.0.0
matplotlib>=3.6.0
plotly>=5.0.0
streamlit>=1.28.0
